pyproject.toml
[project]
name = "zugrafix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.7",
    "faiss-cpu>=1.13.1",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.4",
    "langchain-groq>=1.1.1",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
